refactor(scanner): extract path-accessibility helpers into scanner_paths.go
Move the cohesive path-accessibility / mount-classification cluster —
verifyPathAccessible, classifyStatError, isMountError, testFileAccess,
hasActiveCorruption, LoadActiveCorruptionsForPath — verbatim into
scanner_paths.go (same package). Pure code move: no logic, signature, or
behavior change. scanner.go drops 2070 -> 1969 lines.

Test plan:
- go build ./... ; golangci-lint ./internal/services/... 0 issues
- go test ./internal/services/ pass
